feat(v2): run preflights commands
