+++ cmd/installer/cli/join_runpreflights.go
@@ -3,33 +3,27 @@ package cli
 import (
 	"context"
 	"fmt"
-	"os"
-	"strings"
 
-	ecv1beta1 "github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
 	"github.com/replicatedhq/embedded-cluster/pkg/configutils"
 	"github.com/replicatedhq/embedded-cluster/pkg/kotsadm"
-	"github.com/replicatedhq/embedded-cluster/pkg/netutils"
+	"github.com/replicatedhq/embedded-cluster/pkg/preflights"
 	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
-	"github.com/replicatedhq/embedded-cluster/pkg/versions"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 )
 
 func JoinRunPreflightsCmd(ctx context.Context, name string) *cobra.Command {
-	var (
-		airgapBundle     string
-		networkInterface string
-		assumeYes        bool
-	)
+	var flags Join2CmdFlags
+
 	cmd := &cobra.Command{
 		Use:           "run-preflights",
 		Short:         fmt.Sprintf("Run join host preflights for %s", name),
+		Args:          cobra.ExactArgs(2),
 		SilenceErrors: true,
 		SilenceUsage:  true,
 		PreRunE: func(cmd *cobra.Command, args []string) error {
-			if os.Getuid() != 0 {
-				return fmt.Errorf("run-preflights command must be run as root")
+			if err := preRunJoin2(&flags); err != nil {
+				return err
 			}
 
 			return nil
@@ -38,110 +32,75 @@ func JoinRunPreflightsCmd(ctx context.Context, name string) *cobra.Command {
 			runtimeconfig.Cleanup()
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			if len(args) != 2 {
-				return fmt.Errorf("usage: %s join preflights <url> <token>", name)
-			}
-
 			logrus.Debugf("fetching join token remotely")
-			jcmd, err := kotsadm.GetJoinToken(cmd.Context(), args[0], args[1])
+			jcmd, err := kotsadm.GetJoinToken(ctx, args[0], args[1])
 			if err != nil {
 				return fmt.Errorf("unable to get join token: %w", err)
 			}
-
-			runtimeconfig.Set(jcmd.InstallationSpec.RuntimeConfig)
-			os.Setenv("TMPDIR", runtimeconfig.EmbeddedClusterTmpSubDir())
-
-			// check to make sure the version returned by the join token is the same as the one we are running
-			if strings.TrimPrefix(jcmd.EmbeddedClusterVersion, "v") != strings.TrimPrefix(versions.Version, "v") {
-				return fmt.Errorf("embedded cluster version mismatch - this binary is version %q, but the cluster is running version %q", versions.Version, jcmd.EmbeddedClusterVersion)
-			}
-
-			setProxyEnv(jcmd.InstallationSpec.Proxy)
-
-			networkInterfaceFlag, err := cmd.Flags().GetString("network-interface")
-			if err != nil {
-				return fmt.Errorf("unable to get network-interface flag: %w", err)
-			}
-			proxyOK, localIP, err := checkProxyConfigForLocalIP(jcmd.InstallationSpec.Proxy, networkInterfaceFlag)
-			if err != nil {
-				return fmt.Errorf("failed to check proxy config for local IP: %w", err)
-			}
-			if !proxyOK {
-				return fmt.Errorf("no-proxy config %q does not allow access to local IP %q", jcmd.InstallationSpec.Proxy.NoProxy, localIP)
-			}
-
-			isAirgap := false
-			if airgapBundle != "" {
-				isAirgap = true
-			}
-			logrus.Debugf("materializing binaries")
-			if err := materializeFiles(airgapBundle); err != nil {
+			if err := runJoinRunPreflights(cmd.Context(), cmd, name, flags, jcmd); err != nil {
 				return err
 			}
-
-			if err := configutils.ConfigureSysctl(); err != nil {
-				return err
-			}
-
-			opts := addonsApplierOpts{
-				assumeYes:    assumeYes,
-				license:      "",
-				airgapBundle: airgapBundle,
-				overrides:    "",
-				privateCAs:   nil,
-				configValues: "",
-			}
-			applier, err := getAddonsApplier(cmd, opts, "", jcmd.InstallationSpec.Proxy)
-			if err != nil {
-				return err
-			}
-
-			podCIDR, serviceCIDR, err := netutils.SplitNetworkCIDR(ecv1beta1.DefaultNetworkCIDR)
-			if err != nil {
-				return fmt.Errorf("unable to split default network CIDR: %w", err)
-			}
-
-			if jcmd.InstallationSpec.Network != nil {
-				if jcmd.InstallationSpec.Network.PodCIDR != "" {
-					podCIDR = jcmd.InstallationSpec.Network.PodCIDR
-				}
-				if jcmd.InstallationSpec.Network.ServiceCIDR != "" {
-					serviceCIDR = jcmd.InstallationSpec.Network.ServiceCIDR
-				}
-			}
-
-			cidrCfg := &CIDRConfig{
-				PodCIDR:     podCIDR,
-				ServiceCIDR: serviceCIDR,
-			}
-
-			logrus.Debugf("running host preflights")
-			replicatedAPIURL := jcmd.InstallationSpec.MetricsBaseURL
-			proxyRegistryURL := fmt.Sprintf("https://%s", runtimeconfig.ProxyRegistryAddress)
-			if err := RunHostPreflights(cmd, applier, replicatedAPIURL, proxyRegistryURL, isAirgap, jcmd.InstallationSpec.Proxy, cidrCfg, jcmd.TCPConnectionsRequired, assumeYes); err != nil {
-				if err == ErrPreflightsHaveFail {
-					return ErrNothingElseToAdd
-				}
-				return err
-			}
-
-			logrus.Info("Host preflights completed successfully")
 
 			return nil
 		},
 	}
 
-	cmd.Flags().StringVar(&airgapBundle, "airgap-bundle", "", "Path to the air gap bundle. If set, the installation will complete without internet access.")
-	cmd.Flags().MarkHidden("airgap-bundle")
-	cmd.Flags().StringVar(&networkInterface, "network-interface", "", "The network interface to use for the cluster")
-
-	cmd.Flags().Bool("skip-host-preflights", false, "Skip host preflight checks. This is not recommended and has been deprecated.")
-	cmd.Flags().MarkHidden("skip-host-preflights")
-	cmd.Flags().Bool("ignore-host-preflights", false, "Run host preflight checks, but prompt the user to continue if they fail instead of exiting.")
-	cmd.Flags().MarkHidden("ignore-host-preflights")
-
-	cmd.Flags().BoolVar(&assumeYes, "yes", false, "Assume yes to all prompts.")
-	cmd.Flags().SetNormalizeFunc(normalizeNoPromptToYes)
+	if err := addJoinFlags(cmd, &flags); err != nil {
+		panic(err)
+	}
 
 	return cmd
+}
+
+func runJoinRunPreflights(ctx context.Context, cmd *cobra.Command, name string, flags Join2CmdFlags, jcmd *kotsadm.JoinCommandResponse) error {
+	if err := runJoinVerifyAndPrompt(name, flags, jcmd); err != nil {
+		return err
+	}
+
+	logrus.Debugf("materializing %s binaries", name)
+	if err := materializeFiles(flags.airgapBundle); err != nil {
+		return err
+	}
+
+	logrus.Debugf("configuring sysctl")
+	if err := configutils.ConfigureSysctl(); err != nil {
+		return fmt.Errorf("unable to configure sysctl: %w", err)
+	}
+
+	cidrCfg, err := getJoinCIDRConfig(jcmd)
+	if err != nil {
+		return fmt.Errorf("unable to get join CIDR config: %w", err)
+	}
+
+	logrus.Debugf("running join preflights")
+	if err := runJoinPreflights(ctx, jcmd, flags, cidrCfg); err != nil {
+		return fmt.Errorf("unable to run join preflights: %w", err)
+	}
+
+	logrus.Info("Host preflights completed successfully")
+
+	return nil
+}
+
+func runJoinPreflights(ctx context.Context, jcmd *kotsadm.JoinCommandResponse, flags Join2CmdFlags, cidrCfg *CIDRConfig) error {
+	logrus.Debugf("running host preflights")
+	if err := preflights.PrepareAndRun(ctx, preflights.PrepareAndRunOptions{
+		ReplicatedAPIURL:       jcmd.InstallationSpec.MetricsBaseURL, // MetricsBaseURL is the replicated.app endpoint url
+		ProxyRegistryURL:       fmt.Sprintf("https://%s", runtimeconfig.ProxyRegistryAddress),
+		Proxy:                  jcmd.InstallationSpec.Proxy,
+		PodCIDR:                cidrCfg.PodCIDR,
+		ServiceCIDR:            cidrCfg.ServiceCIDR,
+		IsAirgap:               flags.isAirgap,
+		SkipHostPreflights:     flags.skipHostPreflights,
+		IgnoreHostPreflights:   flags.ignoreHostPreflights,
+		AssumeYes:              flags.assumeYes,
+		TCPConnectionsRequired: jcmd.TCPConnectionsRequired,
+	}); err != nil {
+		if err == preflights.ErrPreflightsHaveFail {
+			return ErrNothingElseToAdd
+		}
+		return fmt.Errorf("unable to prepare and run preflights: %w", err)
+	}
+
+	return nil
 }
